requirements.txt
boto3==1.9.175
coreapi==2.3.3
dj-database-url==0.5.0
Django==2.2.2
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.9.4
factory-boy==2.12.0
gunicorn==19.9.0
matplotlib==3.1.0
psycopg2-binary==2.8.3
requests==2.22.0
